requirements.txt
python-firebase==1.2
Flask==0.10.1
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.1